server/requirements.txt
websockets==14.1
google-auth==2.37.0
certifi==2024.12.14
requests==2.32.3
google-genai==1.3.0
aiohttp==3.11.11
python-dotenv==1.0.1
google-api-python-client==2.122.0
google-auth-oauthlib==1.2.0